scripts: add ondir
It can help automatical task when enter/leaving directory.
I'm using it for setting the root dir, so I can replace cd ../../../../ by cd $ROOT.
Though I think the best way is local directory bashrc, so I don't need to add all the
setting in the .ondirrc and can be commited to project.